pyproject.toml
[project]
name = "feishu-codex"
version = "0.1.0"
description = "Bridge Feishu chats to Codex via the native app-server protocol."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "lark-oapi>=1.4.23",
  "PyYAML>=6.0",
  "websockets>=15.0",
]

[project.scripts]
feishu-codex = "bot.manage_cli:main"
feishu-codexd = "bot.__main__:main"
feishu-codexctl = "bot.feishu_codexctl:main"
fcodex = "bot.fcodex:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["bot*"]
